+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/pablo/PROJECTS/my-app/src/App.tsx":"1","/home/pablo/PROJECTS/my-app/src/index.tsx":"2","/home/pablo/PROJECTS/my-app/src/components/TodoList.tsx":"3","/home/pablo/PROJECTS/my-app/src/components/Todo.tsx":"4"},{"size":1617,"mtime":1606242900293,"results":"5","hashOfConfig":"6"},{"size":153,"mtime":1606225700850,"results":"7","hashOfConfig":"6"},{"size":356,"mtime":1606242618913,"results":"8","hashOfConfig":"6"},{"size":426,"mtime":1606242729169,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1dh6ulw",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/pablo/PROJECTS/my-app/src/App.tsx",[],"/home/pablo/PROJECTS/my-app/src/index.tsx",[],"/home/pablo/PROJECTS/my-app/src/components/TodoList.tsx",[],"/home/pablo/PROJECTS/my-app/src/components/Todo.tsx",[]]
+[{"/home/pablo/PROJECTS/my-app/src/App.tsx":"1","/home/pablo/PROJECTS/my-app/src/index.tsx":"2","/home/pablo/PROJECTS/my-app/src/pages/todos-dashboard/components/Todo.tsx":"3","/home/pablo/PROJECTS/my-app/src/pages/landing-page/components/NavBar.tsx":"4"},{"size":1617,"mtime":1606242900293,"results":"5","hashOfConfig":"6"},{"size":153,"mtime":1606225700850,"results":"7","hashOfConfig":"6"},{"size":439,"mtime":1606251885913,"results":"8","hashOfConfig":"6"},{"size":786,"mtime":1606253463547,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1dh6ulw",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/pablo/PROJECTS/my-app/src/App.tsx",[],"/home/pablo/PROJECTS/my-app/src/index.tsx",[],"/home/pablo/PROJECTS/my-app/src/pages/todos-dashboard/components/Todo.tsx",[],"/home/pablo/PROJECTS/my-app/src/pages/landing-page/components/NavBar.tsx",[]]
