+++ src/App.tsx
@@ -1,60 +1,9 @@
 import React from "react";
-import TodoList from "./components/TodoList";
-import { v4 as uuidv4 } from "uuid";
-import { TodoTask } from "./components/types";
-
-const LOCAL_STORAGE_KEY = "todoApp.todos";
+import NavBar from "./pages/landing-page/components/NavBar";
+import TodosBoard from "./pages/todos-dashboard/TodosBoard";
 
 export function App() {
-  const [todos, setTodos] = React.useState<TodoTask[]>([]);
-  const todoNameRef = React.useRef<HTMLInputElement>(null);
-
-  React.useEffect(() => {
-    const storedTodos = localStorage.getItem(LOCAL_STORAGE_KEY);
-
-    if (storedTodos) {
-      setTodos(JSON.parse(storedTodos));
-    }
-  }, []);
-
-  React.useEffect(() => {
-    localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(todos));
-  }, [todos]);
-
-  function toggleTodo(id: string) {
-    const newTodos = [...todos];
-    const todo = newTodos.find((todo) => todo.id === id);
-    if (todo) todo.complete = !todo.complete;
-    setTodos(newTodos);
-  }
-
-  function handleAddTodo() {
-    if (todoNameRef.current) {
-      const name = todoNameRef.current.value;
-      console.log(name);
-
-      setTodos((prevTodos) => {
-        return [...prevTodos, { id: uuidv4(), name: name, complete: false }];
-      });
-      todoNameRef.current.value = "";
-    }
-    return;
-  }
-
-  function handleClearTodos() {
-    const newTodos = todos.filter((todo) => !todo.complete);
-    setTodos(newTodos);
-  }
-
-  return (
-    <>
-      <TodoList todos={todos} toggleTodo={toggleTodo} />
-      <input ref={todoNameRef} type="text" />
-      <button onClick={handleAddTodo}>Add todo</button>
-      <button onClick={handleClearTodos}>Clear completed todos</button>
-      <div>{todos.filter((todo) => !todo.complete).length}</div>
-    </>
-  );
+  return <NavBar />;
 }
 
 export default App;
