+++ src/pages/landing-page/Dashboard.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import NavBar from "./components/NavBar";
+import "./dashboardStyles.css";
+
+export default function Dashboard() {
+  return (
+    <>
+      <NavBar />
+    </>
+  );
+}
