+++ src/pages/todos-dashboard/components/TodoList.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Todo from "./Todo";
-import { TodoTask } from "./types";
+import { TodoTask } from "../../../types/types";
 
 export default function TodoList({
   todos,
